more stuff

+++ ldm/models/unet.py
@@ -58,7 +58,7 @@ class UNetModel(nn.Module):
             self.conv_resample = conv_resample
             self.num_classes = num_classes
             self.use_checkpoint = use_checkpoint
-            self.dtype = torch.float16
+            self.dtype = torch.float32
             self.num_heads = num_heads
             self.num_head_channels = num_head_channels
             self.num_heads_upsample = num_heads_upsample
